moved py2/3 compatibility code over to py3_compat.py
