+++ TileStache/Sandwich.py
@@ -80,7 +80,7 @@ A complete example configuration might look like this:
       {
         "name": "Test"
       },
-      "layers": 
+      "layers":
       {
         "base":
         {
@@ -118,14 +118,10 @@ A complete example configuration might look like this:
     }
 """
 from re import search
-try:
-    from io import StringIO
-except ImportError:
-    # Python 2
-    from StringIO import StringIO
+from io import BytesIO
 from itertools import product
-from urlparse import urljoin
-from urllib import urlopen
+
+from .py3_compat import urljoin, urlopen, unicode
 
 from . import Core
 
@@ -158,30 +154,27 @@ try:
 except ImportError:
     # Well, this will not work.
     pass
-    
-# only need to check for py3 once: import unicode from .
-from . import unicode
 
 class Provider:
     """ Sandwich Provider.
-    
+
         Stack argument is a list of layer dictionaries described in module docs.
     """
     def __init__(self, layer, stack):
         self.layer = layer
         self.config = layer.config
         self.stack = stack
-    
+
     @staticmethod
     def prepareKeywordArgs(config_dict):
         """ Convert configured parameters to keyword args for __init__().
         """
         return {'stack': config_dict['stack']}
-    
+
     def renderTile(self, width, height, srs, coord):
-        
+
         rendered = self.draw_stack(coord, dict())
-        
+
         if rendered.size() == (width, height):
             return rendered.image()
         else:
@@ -192,14 +185,14 @@ class Provider:
 
             Given a coordinate, return an output image with the results of all the
             layers in this stack pasted on in turn.
-        
+
             Final argument is a dictionary used to temporarily cache results
             of layers retrieved from layer_bitmap(), to speed things up in case
             of repeatedly-used identical images.
         """
         # start with an empty base
         rendered = Blit.Color(0, 0, 0, 0)
-    
+
         for layer in self.stack:
             if 'zoom' in layer and not in_zoom(coord, layer['zoom']):
                 continue
@@ -207,32 +200,32 @@ class Provider:
             #
             # Prepare pixels from elsewhere.
             #
-        
+
             source_name, mask_name, color_name = [layer.get(k, None) for k in ('src', 'mask', 'color')]
-    
+
             if source_name and color_name and mask_name:
                 raise Core.KnownUnknown("You can't specify src, color and mask together in a Sandwich Layer: %s, %s, %s" % (repr(source_name), repr(color_name), repr(mask_name)))
-        
+
             if source_name and source_name not in tiles:
                 if source_name in self.config.layers:
                     tiles[source_name] = layer_bitmap(self.config.layers[source_name], coord)
                 else:
                     tiles[source_name] = local_bitmap(source_name, self.config, coord, self.layer.dim)
-        
+
             if mask_name and mask_name not in tiles:
                 tiles[mask_name] = layer_bitmap(self.config.layers[mask_name], coord)
-        
+
             #
             # Build up the foreground layer.
             #
-        
+
             if source_name and color_name:
                 # color first, then layer
                 foreground = make_color(color_name).blend(tiles[source_name])
-        
+
             elif source_name:
                 foreground = tiles[source_name]
-        
+
             elif color_name:
                 foreground = make_color(color_name)
 
@@ -241,23 +234,23 @@ class Provider:
 
             else:
                 raise Core.KnownUnknown("You have to provide at least some combination of src, color and mask to Sandwich Layer")
-        
+
             #
             # Do the final composition with adjustments and blend modes.
             #
-        
+
             for (name, args) in layer.get('adjustments', []):
                 adjustfunc = adjustment_names.get(name)(*args)
                 foreground = foreground.adjust(adjustfunc)
-        
+
             opacity = float(layer.get('opacity', 1.0))
             blendfunc = blend_modes.get(layer.get('mode', None), None)
-        
+
             if mask_name:
                 rendered = rendered.blend(foreground, tiles[mask_name], opacity, blendfunc)
             else:
                 rendered = rendered.blend(foreground, None, opacity, blendfunc)
-    
+
         return rendered
 
 def local_bitmap(source, config, coord, dim):
@@ -265,53 +258,53 @@ def local_bitmap(source, config, coord, dim):
     """
     address = urljoin(config.dirpath, source)
     bytes = urlopen(address).read()
-    image = Image.open(StringIO(bytes)).convert('RGBA')
-    
+    image = Image.open(BytesIO(bytes)).convert('RGBA')
+
     coord = coord.zoomBy(8)
     w, h, col, row = image.size[0], image.size[1], int(coord.column), int(coord.row)
-    
+
     x = w * (col / w) - col
     y = h * (row / h) - row
-    
+
     output = Image.new('RGBA', (dim, dim))
-    
+
     for (x, y) in product(range(x, dim, w), range(y, dim, h)):
         # crop the top-left if needed
         xmin = 0 if x > 0 else -x
         ymin = 0 if y > 0 else -y
-        
+
         # don't paste up and to the left
         x = x if x >= 0 else 0
         y = y if y >= 0 else 0
-        
+
         output.paste(image.crop((xmin, ymin, w, h)), (x, y))
-    
+
     return Blit.Bitmap(output)
 
 def layer_bitmap(layer, coord):
     """ Return Blit.Bitmap representation of tile from a given layer.
-    
+
         Uses TileStache.getTile(), so caches are read and written as normal.
     """
     from . import getTile
 
     mime, body = getTile(layer, coord, 'png')
-    image = Image.open(StringIO(body)).convert('RGBA')
+    image = Image.open(BytesIO(body)).convert('RGBA')
 
     return Blit.Bitmap(image)
 
 def in_zoom(coord, range):
     """ Return True if the coordinate zoom is within the textual range.
-    
+
         Range might look like "1-10" or just "5".
     """
     zooms = search("^(\d+)-(\d+)$|^(\d+)$", range)
-    
+
     if not zooms:
         raise Core.KnownUnknown("Bad zoom range in a Sandwich Layer: %s" % repr(range))
-    
+
     min_zoom, max_zoom, at_zoom = zooms.groups()
-    
+
     if min_zoom is not None and max_zoom is not None:
         min_zoom, max_zoom = int(min_zoom), int(max_zoom)
 
@@ -320,12 +313,12 @@ def in_zoom(coord, range):
 
     else:
         min_zoom, max_zoom = 0, float('inf')
-    
+
     return min_zoom <= coord.zoom and coord.zoom <= max_zoom
 
 def make_color(color):
     """ Convert colors expressed as HTML-style RGB(A) strings to Blit.Color.
-        
+
         Examples:
           white: "#ffffff", "#fff", "#ffff", "#ffffffff"
           black: "#000000", "#000", "#000f", "#000000ff"
@@ -347,7 +340,7 @@ def make_color(color):
 
     elif len(color) == 5:
         color = ''.join([color[i] for i in (0, 1, 1, 2, 2, 3, 3, 4, 4)])
-    
+
     try:
         r = int(color[1:3], 16)
         g = int(color[3:5], 16)
